removing unused modules and refactoring downloads to use only one module and avoid using spark for small downloads
