+++ occurrence-download/docker/Dockerfile
@@ -4,5 +4,5 @@ FROM docker.gbif.org/gbif/base-image:1.0.0
 ARG JAR_FILE=${JAVA_JAR_FILE}
 USER root
 RUN mkdir /jobs
-COPY --chown=1000:1000 "./target/${JAR_FILE}" "/jobs/occurrence-download-spark.jar"
+COPY --chown=1000:1000 "./target/${JAR_FILE}" "/jobs/occurrence-download.jar"
 USER gbif
