--- occurrence-trino-udf/src/main/java/org/gbif/occurrence/trino/processor/clients/SpeciesWsClient.java
@@ -1,30 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.gbif.occurrence.trino.processor.clients;
-
-import org.gbif.api.model.checklistbank.NameUsage;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-
-@RequestMapping(
-    method = RequestMethod.GET,
-    value = "species",
-    produces = MediaType.APPLICATION_JSON_VALUE)
-public interface SpeciesWsClient {
-
-  @RequestMapping(value = "{key}")
-  NameUsage get(@PathVariable("key") String nubKey);
-}
