--- occurrence-trino-udf/src/main/java/org/gbif/occurrence/trino/udf/GbifUdfs.java
@@ -1,35 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.gbif.occurrence.trino.udf;
-
-import io.trino.spi.Plugin;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
-
-public class GbifUdfs implements Plugin {
-
-  @Override
-  public Set<Class<?>> getFunctions() {
-    return new HashSet<>(
-        Arrays.asList(
-            DataCleanUdfs.class,
-            DateTimeUdfs.class,
-            ContainsUdf.class,
-            BasisOfRecordParseUDF.class,
-            DateParseUDF.class,
-            CoordinateCountryParseUDF.class,
-            SpeciesMatchUDF.class));
-  }
-}
