--- occurrence-trino-udf/src/main/resources/META-INF/services/io.trino.spi.Plugin
@@ -1,1 +1,0 @@
-org.gbif.occurrence.trino.udf.GbifUdfs
