+++ pom.xml
@@ -100,8 +100,6 @@
     <module>occurrence-table-build-spark</module>
     <module>occurrence-download-launcher</module>
     <module>occurrence-spark-udf</module>
-    <module>occurrence-download-spark</module>
-    <module>occurrence-table-build-trino</module>
   </modules>
 
   <properties>
